fix(leads): a lead nobody owns breaches again, unless nobody asked us for it
main is red. #5015 made an ownerless breach reach the configured intake seat;
#5201 added `AND owner_id IS NOT NULL` to the breach scan and left both of that
feature's tests failing without touching them.

The two changes are each right about their own case and the discriminator was
wrong. #5015 seeds an INBOUND nobody picked up — a person waiting, with no one
answerable, which is the queue's worst case. #5201 seeds a SITEREAD — a name a
pass read off a company's website, which is not somebody's work and whose clock
nobody started. Ownership does not tell those apart; it excludes both, which is
why one PR's fix broke the other's tests.

lead_source.intent already draws the line, and says so in its own words:
siteread and crawl are low "for crawl's reason: nobody asked us for anything".
The scan reads that rather than naming those two sources, so a prospecting
source an operator adds gets the rule with it. An owned lead is unconditional —
somebody took it on, so the clock is theirs however it arrived.

Each half of the condition is required by a different PR's test. Dropping the
intent check fails #5201's; reverting to owner-only fails both of #5015's.

WHAT MADE THE EXCLUSION LOOK NECESSARY is closed rather than argued away. Its
comment was right that escalating an unowned row stamps sla_breached_at and
then suppresses the real escalation once a human takes it on — because the
take-on restarts the clock through COALESCE(routed_at, created_at) while the
stamp keeps the row out of the scan for good. So the take-on now clears the
stamp with the same condition that starts the clock, on both paths, and the two
move together.

leadsla.go was at 498 of its 500 ceiling, so the sweep moves to leadslascan.go:
what an SLA state IS on one side, and the pass that acts on it — the half with
the judgement in it — on the other.
